Version 1.0 It kind of works
- Written Files can now be reread and be put back in the system via datamanagement.queueWriter.Stringreader.java
- Fully Implemented Events
- Added Helperclasses for the events and for JSON parsing
- Changed the Database Design by removing Impressions and Engagements from Tweet_Engagements and added Date to User_Interacts_User and User_Interacts_Tweet
- Small bugfixes across the entire project
- Removed unnecessary keys from the KeyManager 

- There is no way to shutdown the Server at the moment since the UI Layer is still not implemented
- Events that can be ignored are not dropped but are written to File as an error in EventParser.addAll()

+++ src/main/java/everyst/analytics/listner/dataManagement/queueWriter/FileManager.java
@@ -55,6 +55,8 @@ public class FileManager {
 	public void moveActiveToOld() {
 		try {
 			// rename the active file to a name which is unique
+			if (!activeFile.exists())
+				return;
 			Files.move(activeFile, new File(root, new Date().getTime() + ""));
 			activeFile.createNewFile();
 		} catch (IOException e) {
